# docs/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Home Repair Video Analyser",
    "description": "Upload a home repair video; get a structured Gemini 2.5 Flash assessment.",
    "version": "0.2.0"
  },
  "paths": {
...
    "/analyse/photos": {
      "post": {
        "tags": [
          "photo_analysis"
        ],
        "summary": "Analyse trade issue from photos",
        "description": "Submit 1\u20135 photographs of a home repair problem. The service assigns each image a triangulation role (Wide Shot, Close-up, Scale/Context \u2026), preprocesses them for token efficiency, and returns a structured Gemini 2.5 Flash diagnosis.",
        "operationId": "analyse_photos_analyse_photos_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PhotoAnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhotoAnalysisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ImageFeedback": {
        "properties": {
          "index": {
            "type": "integer",
            "title": "Index"
          },
          "role": {
            "type": "string",
            "title": "Role",
            "description": "Triangulation role assigned to this image."
          },
          "quality": {
            "type": "string",
            "enum": [
              "ok",
              "blurry",
              "unidentifiable",
              "unsupported"
            ],
            "title": "Quality"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note",
            "description": "Human-readable explanation when quality != 'ok'."
          }
        },
        "type": "object",
        "required": [
          "index",
          "role",
          "quality"
        ],
        "title": "ImageFeedback",
        "description": "Quality assessment for a single submitted image."
      },
...
      "PhotoAnalysisRequest": {
        "properties": {
          "images": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 5,
            "minItems": 1,
            "title": "Images",
            "description": "1\u20135 image sources. Each must be an HTTPS URL or a base64 data URI (data:image/<type>;base64,<data>). Supported formats: JPEG, PNG, WebP."
          },
          "description": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 10,
            "title": "Description",
            "description": "Brief customer description of the problem (10\u20131000 characters)."
          },
          "trade_category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trade Category",
            "description": "Optional trade hint. Must be one of: ['damp', 'electrical', 'general', 'plumbing', 'roofing', 'structural']. Omit if unknown."
          }
        },
        "type": "object",
        "required": [
          "images",
          "description"
        ],
        "title": "PhotoAnalysisRequest"
      },
      "PhotoAnalysisResponse": {
        "properties": {
          "likely_issue": {
            "type": "string",
            "title": "Likely Issue",
            "description": "Concise one-sentence diagnosis. Set to 'INSUFFICIENT_EVIDENCE' when the images do not provide enough detail for a confident diagnosis."
          },
          "urgency_score": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Urgency Score",
            "description": "1 = cosmetic / can wait, 10 = immediate safety risk."
          },
          "required_tools": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Required Tools",
            "description": "Specific tools needed for the repair."
          },
          "estimated_parts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Estimated Parts",
            "description": "Parts or materials identified (with sizes/specs where visible)."
          },
          "image_feedback": {
            "items": {
              "$ref": "#/components/schemas/ImageFeedback"
            },
            "type": "array",
            "title": "Image Feedback",
            "description": "Per-image quality flags from the preprocessing stage."
          },
          "token_usage_estimate": {
            "$ref": "#/components/schemas/TokenUsage",
            "description": "Actual Gemini token counts \u2014 zero when the API does not return them."
          }
        },
        "type": "object",
        "required": [
          "likely_issue",
          "urgency_score",
          "required_tools",
          "estimated_parts",
          "image_feedback",
          "token_usage_estimate"
        ],
        "title": "PhotoAnalysisResponse"
      },
...
      "TokenUsage": {
        "properties": {
          "prompt_tokens": {
            "type": "integer",
            "title": "Prompt Tokens"
          },
          "completion_tokens": {
            "type": "integer",
            "title": "Completion Tokens"
          },
          "total_tokens": {
            "type": "integer",
            "title": "Total Tokens"
          }
        },
        "type": "object",
        "required": [
          "prompt_tokens",
          "completion_tokens",
          "total_tokens"
        ],
        "title": "TokenUsage",
        "description": "Actual token counts from Gemini \u2014 use to track cost efficiency."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}